P1: Reduce -Wsign-conversion warnings (329→248)
Add size_t casts to sizeof expressions in integer arithmetic:
- vlobject.h: Cast 0 to size_t in VLO_LENGTH abort path
- yaep.c: Cast sizeof multiplications in malloc/memset calls
- yaep.c: Cast CHAR_BIT * sizeof expressions in bitset operations
- yaep.c: Cast VLO_EXPAND size arguments with arithmetic expressions

Eliminated 81 sign-conversion warnings. All 128 tests passing.
Valgrind clean (0 errors, 0 leaks).

+++ src/vlobject.h
@@ -165,7 +165,7 @@ typedef struct
 #ifndef NDEBUG
 #define VLO_LENGTH(vlo) ((vlo).vlo_start != NULL\
                          ? (size_t)((vlo).vlo_free - (vlo).vlo_start) /* Cast to size_t to avoid sign/width conversion warnings */\
-                         : (abort (), 0))
+                         : (abort (), (size_t)0))
 #else
 #define VLO_LENGTH(vlo) ((size_t)((vlo).vlo_free - (vlo).vlo_start)) /* Cast to size_t to avoid sign/width conversion warnings */
 #endif /* #ifndef NDEBUG */
